requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.2.3

# Telegram Bot
python-telegram-bot==20.7
httpx==0.25.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Validation & Serialization
phonenumbers==8.13.27
email-validator==2.1.0

# Photo Processing
Pillow==10.1.0
boto3==1.34.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
click==8.1.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Production
gunicorn==21.2.0